Alter Rust bindings to make investigation easier, reproduce issue 63
This patch introduces a PHP script in DEBUG-sfalvo where I am able to reproduce
the OP's issue.  To do this, I needed to make a small adjustment to the Rust
binding code.

After speaking with Khosrow, he recommended adding a few utility methods to the
Expiration class ("We probably need a few methods in there, including
neverExpire() -> bool, serverDefault() -> bool in addition to inSeconds()
(in seconds should be an i64 that can be negative for the two above values.)).

For now, this is just a research branch; DO NOT merge into main.