CLIENT-3777 Fix anomalous getTtl() behavior
This patch greatly simplifies the getTtl() logic in the original PHP
Rust bindings.  It turns out that the Aerospike server itself maintains
the TTL on behalf of the client; there is no need for the PHP client to
maintain CITRUSLEAF_EPOCH-relative timestamps.  This is in line with
both the C and Java clients, thus implying consistent behavior with
Javascript and Python clients as well.

The Expiration interface has been upgraded as well; it is fully backward
compatible with previous code, but now has getters which allows software
to query its current state.

In addition to this change, a number of PHPUnit tests have been added
which exercises the new behavior along with the Expiration type's new
interface.